backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
python-multipart
pydantic
python-dotenv
